Now supporting reverse geocode

+++ src/main/scala/twitter4s/entities/Area.scala
@@ -1,3 +1,3 @@
 package twitter4s.entities
 
-case class Area(coordinates: Seq[Seq[Seq[Float]]], `type`: String)
+case class Area(coordinates: Seq[Seq[Seq[Double]]], `type`: String)
